comment: second i64 LIMIT bind (list_replies) — suite green locally

+++ crates/services/comment/src/infrastructure/persistence/scylla_comment_repository.rs
@@ -464,7 +464,8 @@ impl CommentRepository for ScyllaCommentRepository {
         limit:      i32,
         page_token: Option<&str>,
     ) -> Result<(Vec<CommentSummary>, Option<String>), CommentError> {
-        let limit  = limit.clamp(1, 100) as i64;
+        // i32 for the CQL int32 LIMIT — same fix as list_top_level above.
+        let limit  = limit.clamp(1, 100);
         let token  = decode_page_token(page_token)?;
         let parent = comment_id.as_uuid();
